backend/requirements.txt
django
djangorestframework
qrcode
pillow
pycrypto